Refactor 'utils.py' into Multiple Files
